requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mysql-connector-python==8.0.28
outcome==1.1.0
protobuf==3.19.4
pycparser==2.21
pyOpenSSL==22.0.0
rfc3986==1.5.0
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
